refactor: View 하위 utils 로직을 Utils 폴더로 이동
- readLineAsync.js
- utils.js
